*: 동물 분양 게시글 코드 리팩토링

+++ src/main/java/com/clab/securecoding/type/dto/AnimalAdoptionBoardRequestDto.java
@@ -24,7 +24,7 @@ public class AnimalAdoptionBoardRequestDto {
 
     private Long age;
 
-    private Long weight;
+    private Double weight;
 
     private String vaccine;
