+++ src/main/java/com/clab/securecoding/type/dto/AnimalAdoptionBoardResponseDto.java
@@ -29,7 +29,7 @@ public class AnimalAdoptionBoardResponseDto {
 
     private Long age;
 
-    private Long weight;
+    private Double weight;
 
     private String vaccine;
