+++ src/main/java/com/clab/securecoding/type/entity/AnimalAdoptionBoard.java
@@ -34,7 +34,7 @@ public class AnimalAdoptionBoard {
 
     private Long age;
 
-    private Long weight;
+    private Double weight;
 
     private String vaccine;
